source tree cleanup
