+++ NFSMW_XenonEffects.ini
@@ -10,8 +10,6 @@ Contrails = 1
 BounceParticles = 1
 ; Enables contrails in NIS
 NISContrails = 0
-; Sparks and contrails will recieve shadows
-PassShadowMap = 0
 ; Enables contrails when NOS is activated and does not fade in contrails
 UseCGStyle = 0
 ; Controls the activation speed of contrails (in m/s)
@@ -28,14 +26,16 @@ FadeOutParticles = 0
 ; Min = 0, Max = target game FPS (by default it's 60, this is the same value as SimRate in Widescreen Fix)
 ; Closer the value is to 0, the effects will have less density
 ; Set to 0 to uncap the emission
-ContrailTargetFPS = 30.0
+ContrailTargetFPS = 60.0
 SparkTargetFPS = 60.0
 ; Controls the effect intensity values of contrails.
 ; If UseCGStyle is enabled, the intensity is not calculated and only uses MaxIntensity.
 ; (MaxIntensity on GameCube and Xbox is 1.0)
 ContrailMinIntensity = 0.1
 ContrailMaxIntensity = 0.75
-; Controls the intensity value of sparks & other emitters. Only works if the emitter has the "zContrail" property enabled.
+; Controls the intensity value of sparks & other emitters.
+; Note that values other than 1.0 set alpha between 0 and 42 (16% opacity) as the intensity value is intended for contrails.
+; A value of 1.0 uses alpha value set in the emitter's Attrib/VLT entry instead.
 SparkIntensity = 1.0
 ; Controls the maximum number of generated particles.
 MaxParticles = 2048
