+++ NFSMW_XenonEffects.vcxproj
@@ -57,9 +57,12 @@
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
-      <AdditionalIncludeDirectories>$(SolutionDir);$(DXSDK_DIR)Include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir);$(SolutionDir)\xenoneffects\src;$(SolutionDir)\packages;$(DXSDK_DIR)Include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <ConformanceMode>false</ConformanceMode>
+      <EnableEnhancedInstructionSet>StreamingSIMDExtensions2</EnableEnhancedInstructionSet>
+      <FloatingPointModel>Fast</FloatingPointModel>
+      <FloatingPointExceptions>false</FloatingPointExceptions>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -80,7 +83,10 @@
       <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_USRDLL;_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <RuntimeLibrary>MultiThreadedDLL</RuntimeLibrary>
       <LanguageStandard>stdcpp20</LanguageStandard>
-      <AdditionalIncludeDirectories>$(SolutionDir);$(DXSDK_DIR)Include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir);$(SolutionDir)\xenoneffects\src;$(SolutionDir)\packages;$(DXSDK_DIR)Include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <EnableEnhancedInstructionSet>StreamingSIMDExtensions2</EnableEnhancedInstructionSet>
+      <FloatingPointModel>Fast</FloatingPointModel>
+      <FloatingPointExceptions>false</FloatingPointExceptions>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -93,20 +99,23 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Attrib\Attrib.h" />
-    <ClInclude Include="Attrib\AttribArray.h" />
-    <ClInclude Include="Attrib\AttribAttribute.h" />
-    <ClInclude Include="Attrib\AttribCollection.h" />
-    <ClInclude Include="Attrib\AttribCore.h" />
-    <ClInclude Include="Attrib\AttribHashMap.h" />
-    <ClInclude Include="Attrib\AttribInstance.h" />
-    <ClInclude Include="Attrib\AttribNode.h" />
-    <ClInclude Include="targetver.h" />
-    <ClInclude Include="UMath\UMath.h" />
+    <ClInclude Include="speed\src\GameDefs.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\Attrib.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribArray.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribAttribute.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribCollection.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribCore.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribHashMap.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribInstance.h" />
+    <ClInclude Include="XenonEffects\src\Attrib\AttribNode.h" />
+    <ClInclude Include="XenonEffects\src\Render\xSprites.hpp" />
+    <ClInclude Include="XenonEffects\src\targetver.h" />
+    <ClInclude Include="XenonEffects\src\UMath\UMath.h" />
+    <ClInclude Include="XenonEffects\src\xSpark.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Attrib\Attrib.cpp" />
-    <ClCompile Include="dllmain.cpp">
+    <ClCompile Include="XenonEffects\src\Attrib\Attrib.cpp" />
+    <ClCompile Include="XenonEffects\src\dllmain.cpp">
       <CompileAsManaged Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">false</CompileAsManaged>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
       </PrecompiledHeader>
@@ -114,9 +123,14 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
       </PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="XenonEffects\src\Render\xSprites.cpp" />
+    <ClCompile Include="XenonEffects\src\UMath\UMath.cpp" />
+    <ClCompile Include="XenonEffects\src\xSpark.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="NFSMW_XenonEffects.ini" />
+    <CopyFileToFolders Include="NFSMW_XenonEffects.ini">
+      <FileType>Document</FileType>
+    </CopyFileToFolders>
     <None Include="post-build\postbuild-debug.bat" />
     <None Include="post-build\postbuild.bat" />
     <None Include="README.md" />
