+++ NFSMW_XenonEffects.vcxproj.filters
@@ -1,43 +1,56 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="dllmain.cpp" />
-    <ClCompile Include="Attrib\Attrib.cpp">
+    <ClCompile Include="XenonEffects\src\dllmain.cpp" />
+    <ClCompile Include="XenonEffects\src\Attrib\Attrib.cpp">
       <Filter>Attrib</Filter>
+    </ClCompile>
+    <ClCompile Include="XenonEffects\src\Render\xSprites.cpp">
+      <Filter>Render</Filter>
+    </ClCompile>
+    <ClCompile Include="XenonEffects\src\xSpark.cpp" />
+    <ClCompile Include="XenonEffects\src\UMath\UMath.cpp">
+      <Filter>UMath</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="targetver.h" />
-    <ClInclude Include="UMath\UMath.h">
+    <ClInclude Include="XenonEffects\src\targetver.h" />
+    <ClInclude Include="XenonEffects\src\UMath\UMath.h">
       <Filter>UMath</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\Attrib.h">
+    <ClInclude Include="XenonEffects\src\Attrib\Attrib.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribCollection.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribCollection.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribCore.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribCore.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribHashMap.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribHashMap.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribArray.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribArray.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribInstance.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribInstance.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribNode.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribNode.h">
       <Filter>Attrib</Filter>
     </ClInclude>
-    <ClInclude Include="Attrib\AttribAttribute.h">
+    <ClInclude Include="XenonEffects\src\Attrib\AttribAttribute.h">
       <Filter>Attrib</Filter>
     </ClInclude>
+    <ClInclude Include="speed\src\GameDefs.h">
+      <Filter>Speed</Filter>
+    </ClInclude>
+    <ClInclude Include="XenonEffects\src\Render\xSprites.hpp">
+      <Filter>Render</Filter>
+    </ClInclude>
+    <ClInclude Include="XenonEffects\src\xSpark.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="NFSMW_XenonEffects.ini" />
     <None Include="post-build\postbuild.bat">
       <Filter>post-build</Filter>
     </None>
@@ -62,5 +75,14 @@
     <Filter Include="UMath">
       <UniqueIdentifier>{e1341417-c79a-4863-a372-f573a4360f20}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Speed">
+      <UniqueIdentifier>{da61c712-88df-46e7-b753-02a779af6400}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Render">
+      <UniqueIdentifier>{0f071db2-8e34-41e9-81da-a4bdec41869f}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <CopyFileToFolders Include="NFSMW_XenonEffects.ini" />
   </ItemGroup>
 </Project>
