+++ XenonEffects/src/Attrib/AttribCore.h
